style: encapsulate deprecated interface and its dependencies into a separate file
